Working on Tunig in the 7373 mecanum drive.  Made some adjustments to the drop lift servo.  needs to be tuned in.
Mecanum issues may be fixed either using power changes or weight... Possible a comibination of the two may lead to the best possible results.

Autonomous is have some issues with angles due to strafing and bad positioning. We may need to adjust the alignment turn speed and also need to resolve our issues with strafing.

We also are having issues with rocking even with a 50% power limit on the drive motors.  More weight may be the best solution for this to maintain speed and lower/center our CG